+++ CHANGES.rst
@@ -6,7 +6,8 @@
 0.0.2 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Metric values are now always native strings. See `issue 2
+  <https://github.com/NextThought/nti.fakestatsd/issues/2>`_.
 
 
 0.0.1 (2018-10-12)